pyproject.toml
[project]
name = "dronerl"
version = "1.1.1"
description = "DroneRL - Smart City Drone Delivery using Tabular Q-Learning"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "Adir Elmakais" }]
dependencies = [
    "pygame>=2.5.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.coverage.run]
source = ["src/dronerl"]

[tool.coverage.report]
fail_under = 85

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/dronerl",
    "--cov-fail-under=85",
    "-ra",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.7",
]
